% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{artificial_data_est}
\alias{artificial_data_est}
\title{Generate data using stratified sampling design}
\usage{
artificial_data_est(N, n, n_est)
}
\arguments{
\item{N}{Population size}

\item{n}{Sample size}

\item{n_est}{Number of strata}
}
\value{
List containing sample data, design matrix, weights, and population data
}
\description{
Generate data using stratified sampling design
}
