% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{artificial_data_poi}
\alias{artificial_data_poi}
\title{Generate data using Poisson sampling design}
\usage{
artificial_data_poi(N, n)
}
\arguments{
\item{N}{Population size}

\item{n}{Sample size}
}
\value{
List containing sample data, design matrix, weights, and population data
}
\description{
Generate data using Poisson sampling design
}
