% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{artificial_data_sys}
\alias{artificial_data_sys}
\title{Generate data using systematic sampling design}
\usage{
artificial_data_sys(N, n)
}
\arguments{
\item{N}{Population size}

\item{n}{Sample size}
}
\value{
List containing sample data, design matrix, weights, and population data
}
\description{
Generate data using systematic sampling design
}
