% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo.bqr.svy.R
\name{as_prior_list_per_tau}
\alias{as_prior_list_per_tau}
\title{Normaliza 'prior' a una lista por cuantil}
\usage{
as_prior_list_per_tau(prior, p, names, taus)
}
\description{
Acepta:
\itemize{
\item NULL  -> recicla mo_prior_default para todos los taus
\item función(tau, p, names) -> prior para cada tau
\item un único mo_bqr_prior o lista válida -> se recicla para todos
\item lista de priors (nombrada o de longitud = length(quantile))
}
}
