% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonCrossingBWQR_AL..R
\name{atualizarBETA}
\alias{atualizarBETA}
\title{Update beta coefficients}
\usage{
atualizarBETA(b, B, X, w, sigma, delta2, theta, v, y)
}
\arguments{
\item{b}{Mean vector of the prior.}

\item{B}{Covariance matrix of the prior.}

\item{X}{Design matrix.}

\item{w}{Vector of observation weights.}

\item{sigma}{Scalar scale parameter.}

\item{delta2}{Scalar parameter.}

\item{theta}{Scalar parameter.}

\item{y}{Vector of responses.}
}
\value{
A vector sampled from the posterior of \eqn{\beta}.
}
\description{
Samples \eqn{\beta} from its conditional posterior distribution given the
current values of the latent variables.
}
