% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonCrossingBWQR_AL..R
\name{atualizarV_GIG}
\alias{atualizarV_GIG}
\title{Sample v from Generalized Inverse Gaussian (GIG)}
\usage{
atualizarV_GIG(y, X, w, beta, delta2, theta, sigma, N)
}
\arguments{
\item{y}{Vector of responses.}

\item{X}{Design matrix.}

\item{w}{Vector of observation weights.}

\item{beta}{Vector of regression coefficients.}

\item{delta2}{Scalar parameter.}

\item{theta}{Scalar parameter.}

\item{sigma}{Scalar scale parameter.}

\item{N}{Number of observations.}
}
\value{
A vector of sampled latent variables \code{v}.
}
\description{
Draws samples from the Generalized Inverse Gaussian distribution using the
\code{GIGrvg::rgig} function.
}
