% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{coef.bqr.svy}
\alias{coef.bqr.svy}
\title{Extract coefficients from fitted models}
\usage{
\method{coef}{bqr.svy}(object, ...)
}
\arguments{
\item{object}{A fitted model object}

\item{...}{Additional arguments}
}
\description{
Extract coefficients from fitted models
}
