% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{convergence_check}
\alias{convergence_check}
\title{Check MCMC convergence diagnostics}
\usage{
convergence_check(object, ...)
}
\arguments{
\item{object}{A fitted model object}

\item{...}{Additional arguments}
}
\description{
Check MCMC convergence diagnostics
}
