% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{.autocovariance_vehtari}
\alias{.autocovariance_vehtari}
\title{Autocovariance calculation for Vehtari method}
\usage{
.autocovariance_vehtari(x, max_lag = NULL)
}
\description{
Autocovariance calculation for Vehtari method
}
\keyword{internal}
