% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{.ess_tail_vehtari_single}
\alias{.ess_tail_vehtari_single}
\title{ESS (tails) for single chain — Vehtari et al. (2021)}
\usage{
.ess_tail_vehtari_single(x, prob = c(0.05, 0.95), split_chains = TRUE)
}
\description{
ESS (tails) for single chain — Vehtari et al. (2021)
}
\keyword{internal}
