% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{.ess_vehtari_single}
\alias{.ess_vehtari_single}
\title{ESS (bulk) for single chain — Vehtari et al. (2021)}
\usage{
.ess_vehtari_single(x, split_chains = TRUE)
}
\description{
ESS (bulk) for single chain — Vehtari et al. (2021)
}
\keyword{internal}
