% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{.integrated_time_vehtari}
\alias{.integrated_time_vehtari}
\title{Integrated autocorrelation time for Vehtari method}
\usage{
.integrated_time_vehtari(autocov)
}
\description{
Integrated autocorrelation time for Vehtari method
}
\keyword{internal}
