% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{.rhat_rank}
\alias{.rhat_rank}
\title{Rank-normalized R-hat (Vehtari et al., 2021)}
\usage{
.rhat_rank(x, split_chains = TRUE)
}
\description{
Rank-normalized R-hat (Vehtari et al., 2021)
}
\keyword{internal}
