% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{fitted.bqr.svy}
\alias{fitted.bqr.svy}
\title{Extract fitted values (not implemented for these models)}
\usage{
\method{fitted}{bqr.svy}(object, ...)
}
\arguments{
\item{object}{A fitted model object}

\item{...}{Additional arguments}
}
\description{
Extract fitted values (not implemented for these models)
}
