% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{logLik.bqr.svy}
\alias{logLik.bqr.svy}
\title{Extract log-likelihood (placeholder - not implemented)}
\usage{
\method{logLik}{bqr.svy}(object, ...)
}
\arguments{
\item{object}{A fitted model object}

\item{...}{Additional arguments}
}
\description{
Extract log-likelihood (placeholder - not implemented)
}
