% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonCrossingBWQR_AL..R
\name{NonCrossingBWQR_AL}
\alias{NonCrossingBWQR_AL}
\title{Gibbs Sampler for Non-Crossing BWQR with Asymmetric Laplace Errors}
\usage{
NonCrossingBWQR_AL(
  y,
  X,
  w,
  tau = 0.5,
  n_mcmc = 50000,
  burnin_mcmc = 10000,
  thin_mcmc = 10
)
}
\arguments{
\item{y}{Vector of responses.}

\item{X}{Design matrix.}

\item{w}{Vector of observation weights.}

\item{tau}{Quantile level (e.g., 0.5 for median regression).}

\item{n_mcmc}{Total number of iterations.}

\item{burnin_mcmc}{Number of burn-in iterations.}

\item{thin_mcmc}{Thinning interval.}
}
\value{
A list with posterior samples of \code{beta} and \code{sigma}.
}
\description{
Runs the full Gibbs sampler for Bayesian Weighted Quantile Regression under
asymmetric Laplace errors using the Non-Crossing model.
}
