% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantile.R
\name{plot.bqr.svy}
\alias{plot.bqr.svy}
\title{Plot method for \code{bqr.svy} objects}
\usage{
\method{plot}{bqr.svy}(x, type = c("trace", "intervals", "quantiles"), ...)
}
\arguments{
\item{x}{An object of class \code{bqr.svy}.}

\item{type}{Type of plot: \code{"trace"}, \code{"intervals"}, or \code{"quantiles"}.}

\item{...}{Additional plotting arguments.}
}
\description{
Plot method for \code{bqr.svy} objects
}
