% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantile.R
\name{plot.mo.bqr.svy}
\alias{plot.mo.bqr.svy}
\title{Plot method for \code{mo.bqr.svy} objects}
\usage{
\method{plot}{mo.bqr.svy}(x, type = c("quantiles", "convergence"), ...)
}
\arguments{
\item{x}{An object of class \code{mo.bqr.svy}.}

\item{type}{Type of plot: \code{"quantiles"} or \code{"convergence"}.}

\item{...}{Additional plotting arguments.}
}
\description{
Plot method for \code{mo.bqr.svy} objects
}
