% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantile.R
\name{plot_diagnostics.bqr.svy}
\alias{plot_diagnostics.bqr.svy}
\title{Diagnostic plot method for \code{bqr.svy} objects}
\usage{
plot_diagnostics.bqr.svy(
  x,
  type = c("trace", "intervals", "quantiles"),
  main = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{bqr.svy}.}

\item{type}{Type of plot: \code{"trace"}, \code{"intervals"}, or \code{"quantiles"}.}

\item{main}{Main title for the plot. If \code{NULL}, a default is constructed.}

\item{...}{Additional plotting arguments.}
}
\description{
Diagnostic plot method for \code{bqr.svy} objects
}
\keyword{internal}
