% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantile.R
\name{plot_quantile_with_points.mo.bqr.svy}
\alias{plot_quantile_with_points.mo.bqr.svy}
\title{Plot observed points and predicted quantile regression curve for \code{mo.bqr.svy}}
\usage{
plot_quantile_with_points.mo.bqr.svy(object, data, predictor, main = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{mo.bqr.svy}.}

\item{data}{A data frame containing the variables used in the model.}

\item{predictor}{Character string with the name of the predictor variable.}

\item{main}{Main title for the plot.}

\item{...}{Additional arguments passed to \code{plot_quantile.mo.bqr.svy}.}
}
\description{
Plot observed points and predicted quantile regression curve for \code{mo.bqr.svy}
}
