% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{print.bqr.svy}
\alias{print.bqr.svy}
\title{Print method for bqr.svy objects}
\usage{
\method{print}{bqr.svy}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{"bqr.svy"}.}

\item{digits}{Integer, number of significant digits to print.}

\item{...}{Additional arguments passed to \code{\link[base]{print}}.}
}
\description{
Print method for bqr.svy objects
}
