% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{print.bwqr_fit}
\alias{print.bwqr_fit}
\title{Print method for bwqr_fit objects}
\usage{
\method{print}{bwqr_fit}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{"bwqr_fit"}.}

\item{digits}{Integer, number of significant digits to print.}

\item{...}{Additional arguments passed to \code{\link[base]{print}}.}
}
\description{
Print method for bwqr_fit objects
}
