% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo.bqr.svy.R
\name{print.mo.bqr.svy}
\alias{print.mo.bqr.svy}
\title{Print method for mo.bqr.svy objects}
\usage{
\method{print}{mo.bqr.svy}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"mo.bqr.svy"}.}

\item{...}{Additional arguments (not used).
Print method for mo.bqr.svy objects}
}
\description{
Displays the fitted call, quantiles, and estimated coefficients per direction.
}
