% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo.bqr.svy.R
\name{print.mo_bqr_prior}
\alias{print.mo_bqr_prior}
\title{Print method for \code{mo_bqr_prior} objects}
\usage{
\method{print}{mo_bqr_prior}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mo_bqr_prior}.}

\item{...}{Additional arguments passed to or from other methods
(currently unused).}
}
\value{
The object \code{x}, invisibly.
}
\description{
This function defines the \code{print()} method for objects of class
\code{mo_bqr_prior}. It displays a summary of the prior structure,
including the length of \code{beta_mean}, the dimension of
\code{beta_cov}, and the values of \code{sigma_shape} and \code{sigma_rate}.
}
\examples{
prior <- mo_prior_default(p = 3)
print(prior)

}
