% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{print.summary.bqr.svy}
\alias{print.summary.bqr.svy}
\title{Print method for summary.bqr.svy objects}
\usage{
\method{print}{summary.bqr.svy}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.bqr.svy"}.}

\item{...}{Additional arguments passed to \code{\link[base]{print}}.}
}
\description{
Print method for summary.bqr.svy objects
}
