% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{print.summary.bwqr_fit}
\alias{print.summary.bwqr_fit}
\title{Print method for summary.bwqr_fit objects}
\usage{
\method{print}{summary.bwqr_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.bwqr_fit"}.}

\item{...}{Additional arguments passed to \code{\link[base]{print}}.}
}
\description{
Print method for summary.bwqr_fit objects
}
