% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{print.summary.mo.bqr.svy}
\alias{print.summary.mo.bqr.svy}
\title{Print method for summary.mo.bqr.svy}
\usage{
\method{print}{summary.mo.bqr.svy}(x, ...)
}
\arguments{
\item{x}{A summary.mo.bqr.svy object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for summary.mo.bqr.svy
}
