% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{print.summary.mo_bqr.svy}
\alias{print.summary.mo_bqr.svy}
\title{Print method for summary.mo_bqr.svy objects}
\usage{
\method{print}{summary.mo_bqr.svy}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.mo_bqr.svy"}.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print method for summary.mo_bqr.svy objects
}
