% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{residuals.bqr.svy}
\alias{residuals.bqr.svy}
\title{Extract residuals (not implemented for these models)}
\usage{
\method{residuals}{bqr.svy}(object, ...)
}
\arguments{
\item{object}{A fitted model object}

\item{...}{Additional arguments}
}
\description{
Extract residuals (not implemented for these models)
}
