% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{summarise_draws_custom}
\alias{summarise_draws_custom}
\title{Summaries and diagnostics for MCMC draws (Vehtari et al., 2021)}
\usage{
summarise_draws_custom(draws, max_lag = 200)
}
\description{
Summaries and diagnostics for MCMC draws (Vehtari et al., 2021)
}
\keyword{internal}
