% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{summary.mo.bqr.svy}
\alias{summary.mo.bqr.svy}
\title{Resumen para objetos mo.bqr.svy (EM)}
\usage{
\method{summary}{mo.bqr.svy}(object, digits = 3, ...)
}
\arguments{
\item{object}{Objeto \code{mo.bqr.svy}.}

\item{digits}{Decimales.}

\item{...}{Ignorado.}
}
\value{
Objeto \code{summary.mo_bqr.svy} (data.frame con atributos).
}
\description{
Resumen para objetos mo.bqr.svy (EM)
}
