% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{summary_bqr}
\alias{summary_bqr}
\title{Helper function for user convenience - unified summary interface}
\usage{
summary_bqr(object, ...)
}
\arguments{
\item{object}{A fitted model object (bqr.svy or mo.bqr.svy)}

\item{...}{Additional arguments passed to specific summary methods}
}
\description{
Helper function for user convenience - unified summary interface
}
