% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{as_mo_bqr_prior}
\alias{as_mo_bqr_prior}
\title{Coerce to a \code{mo_bqr_prior} object}
\usage{
as_mo_bqr_prior(x, ...)
}
\arguments{
\item{x}{Object to coerce.}

\item{...}{Additional arguments (ignored).}
}
\value{
A \code{mo_bqr_prior} object.
}
\description{
Allows conversion from lists or other objects to \code{mo_bqr_prior}.
}
\keyword{internal}
