% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo.bqr.svy.R
\name{mo.bqr.svy}
\alias{mo.bqr.svy}
\title{Multiple-Output Bayesian Quantile Regression for Complex Surveys (EM)}
\usage{
mo.bqr.svy(
  formula,
  weights = NULL,
  data,
  quantile = 0.5,
  algorithm = "em",
  prior = NULL,
  n_dir = 1,
  epsilon = 1e-06,
  max_iter = 1000,
  verbose = FALSE,
  niter = NULL,
  burnin = NULL,
  thin = NULL,
  ...
)
}
\description{
Multiple-Output Bayesian Quantile Regression for Complex Surveys (EM)
}
