% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{plot.bqr.svy}
\alias{plot.bqr.svy}
\title{Plot method for bqr.svy objects}
\usage{
\method{plot}{bqr.svy}(x, type = c("trace", "density", "intervals"), pars = NULL, ...)
}
\arguments{
\item{x}{An object of class 'bqr.svy'}

\item{type}{Type of plot: "trace", "density", "intervals" (default: "trace")}

\item{pars}{Parameters to plot (default: all)}

\item{...}{Additional arguments passed to plotting functions}
}
\description{
Plot method for bqr.svy objects
}
