% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{plot.mo.bqr.svy}
\alias{plot.mo.bqr.svy}
\title{Plot method for mo.bqr.svy objects}
\usage{
\method{plot}{mo.bqr.svy}(x, type = c("quantiles", "convergence"), ...)
}
\arguments{
\item{x}{An object of class 'mo.bqr.svy'}

\item{type}{Type of plot: "quantiles", "convergence" (default: "quantiles")}

\item{...}{Additional arguments passed to plotting functions}
}
\description{
Plot method for mo.bqr.svy objects
}
