% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\docType{methods}
\name{print.bayesQRsurvey}
\alias{print.bayesQRsurvey}
\alias{print.bqr.svy}
\alias{print.mo.bqr.svy}
\title{Print methods for bayesQRsurvey model objects}
\usage{
\method{print}{bqr.svy}(x, digits = 3, ...)

\method{print}{mo.bqr.svy}(x, digits = 3, ...)
}
\description{
print.bayesQRsurvey is an S3 method that prints the content of an S3 object of class
\code{bqr.svy} or \code{mo.bqr.svy} to the console.
}
