% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{print.bqr.svy}
\alias{print.bqr.svy}
\title{Print method for bqr.svy objects}
\usage{
\method{print}{bqr.svy}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class 'bqr.svy'}

\item{digits}{Number of decimal places for output (default: 3)}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for bqr.svy objects
}
