% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo.bqr.svy.R, R/summary_bqr_svy.R
\name{print.mo.bqr.svy}
\alias{print.mo.bqr.svy}
\title{Print method for mo.bqr.svy objects}
\usage{
\method{print}{mo.bqr.svy}(x, digits = 3, ...)

\method{print}{mo.bqr.svy}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{"mo.bqr.svy"}.}

\item{digits}{Integer, number of significant digits to print.}

\item{...}{Additional arguments passed to \code{\link[base]{print}}.}
}
\description{
Print method for mo.bqr.svy objects

Print method for mo.bqr.svy objects
}
