% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{print.summary.bqr.svy}
\alias{print.summary.bqr.svy}
\title{Impresi??n del resumen para objetos summary.bqr.svy}
\usage{
\method{print}{summary.bqr.svy}(x, ...)
}
\description{
Impresi??n del resumen para objetos summary.bqr.svy
}
