% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{print.summary.bwqr_fit}
\alias{print.summary.bwqr_fit}
\title{Impresi??n del resumen para objetos summary.bwqr_fit}
\usage{
\method{print}{summary.bwqr_fit}(x, ...)
}
\description{
Impresi??n del resumen para objetos summary.bwqr_fit
}
