% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo.bqr.svy.R
\name{simulate_mo_bqr_data}
\alias{simulate_mo_bqr_data}
\title{Simulate data for Multiple-Output Bayesian Quantile Regression}
\usage{
simulate_mo_bqr_data(n = 100, p = 2, beta_true = NULL, seed = NULL)
}
\description{
Simulate data for Multiple-Output Bayesian Quantile Regression
}
