% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{summarise_draws_custom}
\alias{summarise_draws_custom}
\title{Compute summary statistics for MCMC draws}
\usage{
summarise_draws_custom(draws, max_lag = 1000)
}
\arguments{
\item{draws}{Matrix of MCMC draws (samples x parameters)}

\item{max_lag}{Maximum lag for effective sample size calculation}
}
\description{
Compute summary statistics for MCMC draws
}
\keyword{internal}
