% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{summary.bqr.svy}
\alias{summary.bqr.svy}
\title{Summary method for bqr.svy objects}
\usage{
\method{summary}{bqr.svy}(object, probs = c(0.025, 0.975), digits = 3, ...)
}
\arguments{
\item{object}{An object of class \code{"bqr.svy"}.}

\item{probs}{Numeric vector of probabilities for summary statistics.}

\item{digits}{Integer, number of significant digits to display.}

\item{...}{Additional arguments passed to other methods.}
}
\description{
Summary method for bqr.svy objects
}
