% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{summary.bwqr_fit}
\alias{summary.bwqr_fit}
\title{Resumen para objetos bwqr_fit (un cuantil)}
\usage{
\method{summary}{bwqr_fit}(object, probs = c(0.025, 0.975), digits = 3, max_lag = 200, ...)
}
\arguments{
\item{object}{Objeto \code{bwqr_fit}.}

\item{probs}{Probabilidades para IC.}

\item{digits}{Decimales para impresi??n.}

\item{max_lag}{M??ximo rezago para autocorrelaci??n (si lo usa summarise_draws_custom).}

\item{...}{Ignorado.}
}
\value{
Objeto de clase \code{summary.bwqr_fit}.
}
\description{
Resumen para objetos bwqr_fit (un cuantil)
}
