% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{summary.bwqr_fit}
\alias{summary.bwqr_fit}
\title{Summary method for bwqr_fit objects}
\usage{
\method{summary}{bwqr_fit}(object, probs = c(0.025, 0.975), digits = 3, max_lag = 200, ...)
}
\arguments{
\item{object}{An object of class \code{"bwqr_fit"}.}

\item{probs}{Numeric vector of probabilities for summary statistics.}

\item{digits}{Integer, number of significant digits to display.}

\item{max_lag}{Integer, maximum lag for autocorrelation diagnostics.}

\item{...}{Additional arguments passed to other methods.}
}
\description{
Summary method for bwqr_fit objects
}
