% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bqr_svy.R
\name{summary_bqr}
\alias{summary_bqr}
\title{Generic summary function for Bayesian Quantile Regression objects}
\usage{
summary_bqr(object, ...)
}
\arguments{
\item{object}{An object containing model fit results.}

\item{...}{Additional arguments passed to other methods.}
}
\description{
Generic summary function for Bayesian Quantile Regression objects
}
